+++ benchmarks/benchmarks/ci/__init__.py
@@ -1,1 +1,0 @@
-"""Quick running benchmarks to be included in iris-esmf-regrid's CI."""