+++ benchmarks/benchmarks/esmf_regridder/__init__.py
@@ -1,4 +1,4 @@
-"""Quick running benchmarks for :mod:`esmf_regrid.esmf_regridder`."""
+"""Benchmarks for :mod:`esmf_regrid.esmf_regridder`."""
 
 import os
 from pathlib import Path
@@ -9,7 +9,6 @@ import iris
 from iris.cube import Cube
 from iris.experimental.ugrid import PARSE_UGRID_ON_LOAD
 
-from benchmarks import disable_repeat_between_setup
 from esmf_regrid.esmf_regridder import GridInfo
 from esmf_regrid.experimental.unstructured_scheme import (
     GridToMeshESMFRegridder,
@@ -124,7 +123,6 @@ class TimeRegridding(MultiGridCompare):
         _ = self.regridder(self.src)
 
 
-@disable_repeat_between_setup
 class TimeLazyRegridding:
     def setup_cache(self):
         SYNTH_DATA_DIR = Path().cwd() / "tmp_data"
@@ -172,8 +170,10 @@ class TimeLazyRegridding:
         _ = regridder(self.src)
 
     def time_regridding_realisation(self, cache):
+        # Don't touch result.data - permanent realisation plays badly with
+        #  ASV's re-run strategy.
         assert self.result.has_lazy_data()
-        _ = self.result.data
+        self.result.core_data().compute()
 
 
 class TimeMeshToGridRegridding(TimeRegridding):
@@ -208,7 +208,6 @@ class TimeMeshToGridRegridding(TimeRegridding):
         self.tgt = tgt
 
 
-@disable_repeat_between_setup
 class TimeLazyMeshToGridRegridding:
     def setup_cache(self):
         SYNTH_DATA_DIR = Path().cwd() / "tmp_data"
@@ -252,8 +251,10 @@ class TimeLazyMeshToGridRegridding:
         _ = regridder(self.src)
 
     def time_regridding_realisation(self, cache):
+        # Don't touch result.data - permanent realisation plays badly with
+        #  ASV's re-run strategy.
         assert self.result.has_lazy_data()
-        _ = self.result.data
+        self.result.core_data().compute()
 
 
 class TimeGridToMeshRegridding(TimeRegridding):
@@ -288,7 +289,6 @@ class TimeGridToMeshRegridding(TimeRegridding):
         self.tgt = tgt
 
 
-@disable_repeat_between_setup
 class TimeLazyGridToMeshRegridding:
     def setup_cache(self):
         SYNTH_DATA_DIR = Path().cwd() / "tmp_data"
@@ -328,8 +328,10 @@ class TimeLazyGridToMeshRegridding:
         _ = regridder(self.src)
 
     def time_regridding_realisation(self, cache):
+        # Don't touch result.data - permanent realisation plays badly with
+        #  ASV's re-run strategy.
         assert self.result.has_lazy_data()
-        _ = self.result.data
+        self.result.core_data().compute()
 
 
 class TimeRegridderIO(MultiGridCompare):