+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - [PR#217](https://github.com/SciTools-incubator/iris-esmf-regrid/pull/217)
   Changed the behaviour of coordinate fetching to allow Cubes with both
   1D DimCoords and 2D AuxCoords. In this case the DimCoords are prioritised.
+- [PR#220](https://github.com/SciTools-incubator/iris-esmf-regrid/pull/220)
+  Matured the benchmarking architecture in line with the latest setup in
+  SciTools/iris.
 
 ## [0.5] - 2022-10-14